refactor(core): Move `typeorm` operators from `PruningService` to `ExecutionRepository` (no-changelog)
